% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_split.R
\name{find_optimal_split}
\alias{find_optimal_split}
\title{A function to find the optimal split ratio along with the split mask for a given dataset and model}
\usage{
find_optimal_split(dataset.name, dataset, model.relation, min_test_size)
}
\arguments{
\item{dataset.name}{Name of the Dataset}

\item{dataset}{Dataset as a data.frame}

\item{model.relation}{Regression model in R format}

\item{min_test_size}{a numeric value in \link{0.0, 1.0} denoting the minimum percentage of entries in test set, by default 0.1}
}
\value{
\enumerate{
\item Split Ratio/Percentage.
\item Optimal Dataset split with optimal split ratio/percentage.
\item Supporting Plots of Mahalanobis Distance vs. Akaike Information Criterion.
}
}
\description{
A function to find the optimal split ratio along with the split mask for a given dataset and model
}
\examples{

# ------------ Example 1 ------------
# data preparation
dataset.name <- "Abalone"
data(abalone)

# defining model relation based on variables of data
model.relation <- WholeWeight ~ Height + LongestShell + Diameter

# function call
find_optimal_split(dataset.name, abalone, model.relation, min_test_size)

}
