% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{get_RSS}
\alias{get_RSS}
\title{Get the Residual Sum of Squares}
\usage{
get_RSS(actual, prediction)
}
\arguments{
\item{actual}{Actual values of response variable}

\item{prediction}{Model predictions of response variable}
}
\value{
A numeric value denoting the residual sum of squares
}
\description{
Get the Residual Sum of Squares
}
\examples{

}
