% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_simulation.R
\name{visualize_simulation}
\alias{visualize_simulation}
\title{Title}
\usage{
visualize_simulation(
  df.scores,
  dataset.name,
  initial.scores,
  model.relation,
  save.plots = TRUE,
  output.dir,
  metric.performance = "Normalized AIC"
)
}
\arguments{
\item{df.scores}{Dataframe consisting of metric performance scores for different train-test splits}

\item{dataset.name}{Name of the Dataset (String)}

\item{initial.scores}{The performance metric value for the given initial train-test split}

\item{model.relation}{The relation used for regression model}

\item{save.plots}{Saves plots in output.dir when set to TRUE}

\item{output.dir}{The path to output directory the plots are saved to}

\item{metric.performance}{The performance metric, usually Normalized AIC}
}
\value{

}
\description{
Title
}
