% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose the Random Split}
\usage{
diagnose(
  dataset.name,
  num.iterations,
  df,
  split.percentage,
  response.var,
  model.relation,
  initial.scores,
  dir
)
}
\arguments{
\item{dataset.name}{Name of the dataset (string)}

\item{model.relation}{The Regression Model to be fitted on the training data}

\item{df.train}{Train set corresponding to the initial split performed by the user}

\item{df.test}{Test set corresponding to the initial split performed by the user}
}
\value{
\enumerate{
\item Decision regarding the randomness of the split.
\item Decision regarding the performance of the model.
\item Supporting Plots of Mahalanobis Distance vs. Akaike Information Criterion.
}
}
\description{
Diagnose the Random Split
}
\examples{

# ---------------- Example 1 ----------------

# data preparation
dataset.name <- "Abalone"
data(abalone)

# intial random split of data
s <- sample(x = 1:nrow(abalone), size = floor(nrow(abalone)*0.7), replace = F)
df.train <- abalone[s, ]
df.test <- abalone[-s, ]

# defining model relation based on variables of data
model.relation <- WholeWeight ~ Height + LongestShell + Diameter

# function call
diagnose(dataset.name, df.train, df.test, model.relation)

# ---------------- Example 2 ----------------

# data preparation
dataset.name <- "Diamonds"
data(diamonds)

# intial random split of data
s <- sample(x = 1:nrow(diamonds), size = floor(nrow(diamonds)*0.7), replace = F)
df.train <- diamonds[s, ]
df.test <- diamonds[-s, ]

# defining model relation based on variables of data
model.relation <- price ~ x:y:z + depth

# function call
diagnose(dataset.name, df.train, df.test, model.relation)


}
