% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose the Random Split}
\usage{
diagnose(dataset_name = "DATASET", df_train, df_test, model_relation)
}
\arguments{
\item{dataset_name}{Name of the dataset (string)}

\item{df_train}{Train set corresponding to the initial split performed by the user}

\item{df_test}{Test set corresponding to the initial split performed by the user}

\item{model_relation}{The Regression Model to be fitted on the training data}
}
\value{
\enumerate{
\item Decision regarding the randomness of the split.
\item Decision regarding the performance of the model.
\item Supporting Plots of Mahalanobis Distance vs. Akaike Information Criterion.
}
}
\description{
Diagnose the Random Split
}
\examples{

# ---------------- Example 1 ----------------

# data preparation
dataset_name <- "Abalone"
data(abalone)

# intial random split of data
s <- sample(x = 1:nrow(abalone), size = floor(nrow(abalone)*0.7), replace = F)
df_train <- abalone[s, ]
df_test <- abalone[-s, ]

# defining model relation based on variables of data
model_relation <- WholeWeight ~ Height + LongestShell + Diameter

# function call
diagnose(dataset_name, df_train, df_test, model_relation)

# ---------------- Example 2 ----------------

# data preparation
dataset_name <- "Diamonds"
data(diamonds)

# intial random split of data
s <- sample(x = 1:nrow(diamonds), size = floor(nrow(diamonds)*0.7), replace = F)
df_train <- diamonds[s, ]
df_test <- diamonds[-s, ]

# defining model relation based on variables of data
model_relation <- price ~ x:z + x + z + depth

# function call
diagnose(dataset_name, df_train, df_test, model_relation)


}
