% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mahalanobis_distance}
\alias{get_mahalanobis_distance}
\title{Obtain a Modified version of the Mahalnobis Distance between two populations}
\usage{
get_mahalanobis_distance(df1, df2)
}
\arguments{
\item{df1}{Population Data Frame from which distance is to be measured}

\item{df2}{Population Data Frame for which distance is to be measured (variables should be same as df1)}
}
\value{
A numerical value denoting the modified mahalanobis distance
}
\description{
Obtain a Modified version of the Mahalnobis Distance between two populations
}
