% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Title}
\usage{
simulate(
  dataset.name,
  df.train,
  df.test,
  num.simulations,
  model.relation,
  split.percentage,
  initial.scores,
  alpha,
  save.plots = TRUE,
  output.dir,
  metric.performance = "Normalized AIC"
)
}
\arguments{
\item{dataset.name}{Name of the Dataset (String)}

\item{df.train}{Train Partition (R DataFrame)}

\item{df.test}{Test Partition (R DataFrame)}

\item{num.simulations}{Number of simulations, defaults to 200}

\item{model.relation}{The relation used for regression model}

\item{split.percentage}{The train-test split percentage}

\item{alpha}{The level of the test for visualize_threshold, default set to 0.05}

\item{save.plots}{Saves plots in output.dir when set to TRUE}

\item{output.dir}{The output directory the plots are saved to}

\item{metric.performance}{The performance metric, usually Normalized AIC}
}
\value{

}
\description{
Title
}
